I declared the properties to reference the views and assigned them to objects. I used the views walkthrough as an example
I was unable to get the program to work during class but I am pushing for partial credit if I cant before the deadline.

I made a last minute adjustment but was unable to test because the emulator refused to update, displaying "the program is unresponsive", however I'm pretty sure this should work.